refactor(api): Separate Game and GameInstance
Move GameInstance into its own file, gameInstance.ts
