+++ src/error.ts
@@ -1,0 +1,15 @@
+export enum ErrorCode {
+    OK,
+    NOT_YET_IMPLEMENTED,
+    INVALID_INPUT,
+    INVALID_STATE
+}
+
+export class RegalError extends Error {
+    code: ErrorCode;
+
+    constructor(message: string = "") {
+        super(`RegalError: ${message}`);
+        Object.setPrototypeOf(this, new.target.prototype);
+    }
+}
